+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.7] - 2025-05-21
+
 ### Added
 
 - Option to pass `trainer_kwargs` to `Gtfs2VecEmbedder`'s `fit` and `fit_transform` functions
@@ -410,7 +412,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Intersection Joiner
 - Geoparquet Loader
 
-[Unreleased]: https://github.com/kraina-ai/srai/compare/0.9.6...HEAD
+[Unreleased]: https://github.com/kraina-ai/srai/compare/0.9.7...HEAD
+
+[0.9.7]: https://github.com/kraina-ai/srai/compare/0.9.6...0.9.7
 
 [0.9.6]: https://github.com/kraina-ai/srai/compare/0.9.5...0.9.6